requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.0
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-model-utils==3.1.2
django-notifications-hq==1.5.0
idna==2.8
jsonfield==2.0.2
mysqlclient==1.4.2
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2017.3
requests==2.21.0
soupsieve==1.8
urllib3==1.24.1
